requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.62.1
freetype-py==2.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.55.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
html5lib==1.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
pyparsing==3.3.2
pypdf==6.9.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.48
svglib==1.6.0
tinycss2==1.5.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.6
xhtml2pdf==0.2.17
gunicorn==21.2.0
